refactor: split telemetry collectors and provider helpers
