+++ internal/providers/ollama/ollama.go
@@ -1,20 +1,15 @@
 package ollama
 
 import (
-	"bufio"
-	"bytes"
 	"context"
 	"database/sql"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"net/http"
 	"net/url"
 	"os"
-	"path/filepath"
 	"regexp"
-	"runtime"
 	"sort"
 	"strconv"
 	"strings"
@@ -1164,89 +1159,6 @@ func resolveCloudBaseURL(acct core.AccountConfig) string {
 	return normalize(defaultCloudBaseURL)
 }
 
-func resolveDesktopDBPath(acct core.AccountConfig) string {
-	if acct.ExtraData != nil {
-		for _, key := range []string{"db_path", "app_db"} {
-			if v := strings.TrimSpace(acct.ExtraData[key]); v != "" {
-				return v
-			}
-		}
-	}
-
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return ""
-	}
-
-	switch runtime.GOOS {
-	case "darwin":
-		return filepath.Join(home, "Library", "Application Support", "Ollama", "db.sqlite")
-	case "linux":
-		candidates := []string{
-			filepath.Join(home, ".local", "share", "Ollama", "db.sqlite"),
-			filepath.Join(home, ".config", "Ollama", "db.sqlite"),
-		}
-		for _, c := range candidates {
-			if fileExists(c) {
-				return c
-			}
-		}
-		return candidates[0]
-	case "windows":
-		appData := os.Getenv("APPDATA")
-		if appData != "" {
-			return filepath.Join(appData, "Ollama", "db.sqlite")
-		}
-		return filepath.Join(home, "AppData", "Roaming", "Ollama", "db.sqlite")
-	default:
-		return filepath.Join(home, ".ollama", "db.sqlite")
-	}
-}
-
-func resolveServerConfigPath(acct core.AccountConfig) string {
-	if acct.ExtraData != nil {
-		if v := strings.TrimSpace(acct.ExtraData["server_config"]); v != "" {
-			return v
-		}
-		if configDir := strings.TrimSpace(acct.ExtraData["config_dir"]); configDir != "" {
-			return filepath.Join(configDir, "server.json")
-		}
-	}
-
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return ""
-	}
-	return filepath.Join(home, ".ollama", "server.json")
-}
-
-func resolveServerLogFiles(acct core.AccountConfig) []string {
-	logDir := ""
-	if acct.ExtraData != nil {
-		logDir = strings.TrimSpace(acct.ExtraData["logs_dir"])
-		if logDir == "" {
-			if configDir := strings.TrimSpace(acct.ExtraData["config_dir"]); configDir != "" {
-				logDir = filepath.Join(configDir, "logs")
-			}
-		}
-	}
-	if logDir == "" {
-		home, err := os.UserHomeDir()
-		if err != nil {
-			return nil
-		}
-		logDir = filepath.Join(home, ".ollama", "logs")
-	}
-
-	pattern := filepath.Join(logDir, "server*.log")
-	files, err := filepath.Glob(pattern)
-	if err != nil {
-		return nil
-	}
-	sort.Strings(files)
-	return files
-}
-
 func queryCount(ctx context.Context, db *sql.DB, query string) (int64, error) {
 	var count int64
 	if err := db.QueryRowContext(ctx, query).Scan(&count); err != nil {
@@ -2053,506 +1965,6 @@ func populateDailySeriesFromDB(ctx context.Context, db *sql.DB, snap *core.Usage
 	}
 
 	return nil
-}
-
-func parseLogFile(path string, onEvent func(ginLogEvent)) error {
-	f, err := os.Open(path)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	scanner := bufio.NewScanner(f)
-	const maxLogLine = 1024 * 1024
-	buf := make([]byte, 0, 64*1024)
-	scanner.Buffer(buf, maxLogLine)
-
-	for scanner.Scan() {
-		line := scanner.Text()
-		event, ok := parseGINLogLine(line)
-		if !ok {
-			continue
-		}
-		onEvent(event)
-	}
-	if err := scanner.Err(); err != nil {
-		return err
-	}
-	return nil
-}
-
-func parseGINLogLine(line string) (ginLogEvent, bool) {
-	line = strings.TrimSpace(line)
-	if !strings.HasPrefix(line, "[GIN]") {
-		return ginLogEvent{}, false
-	}
-
-	parts := strings.Split(line, "|")
-	if len(parts) < 5 {
-		return ginLogEvent{}, false
-	}
-
-	left := strings.TrimSpace(strings.TrimPrefix(parts[0], "[GIN]"))
-	leftParts := strings.Split(left, " - ")
-	if len(leftParts) != 2 {
-		return ginLogEvent{}, false
-	}
-
-	timestamp, err := time.ParseInLocation("2006/01/02 15:04:05", strings.TrimSpace(leftParts[0])+" "+strings.TrimSpace(leftParts[1]), time.Local)
-	if err != nil {
-		return ginLogEvent{}, false
-	}
-
-	status, err := strconv.Atoi(strings.TrimSpace(parts[1]))
-	if err != nil {
-		return ginLogEvent{}, false
-	}
-
-	durationText := strings.TrimSpace(parts[2])
-	durationText = strings.ReplaceAll(durationText, "µ", "u")
-	duration, err := time.ParseDuration(durationText)
-	if err != nil {
-		return ginLogEvent{}, false
-	}
-
-	methodPath := strings.TrimSpace(parts[4])
-	methodPathParts := strings.Fields(methodPath)
-	if len(methodPathParts) < 2 {
-		return ginLogEvent{}, false
-	}
-
-	method := strings.TrimSpace(methodPathParts[0])
-	path := strings.Trim(strings.TrimSpace(methodPathParts[1]), `"`)
-	if method == "" || path == "" {
-		return ginLogEvent{}, false
-	}
-
-	return ginLogEvent{
-		Timestamp: timestamp,
-		Status:    status,
-		Duration:  duration,
-		Method:    method,
-		Path:      path,
-	}, true
-}
-
-func isInferencePath(path string) bool {
-	switch path {
-	case "/api/chat", "/api/generate", "/api/embed", "/api/embeddings",
-		"/v1/chat/completions", "/v1/completions", "/v1/responses", "/v1/embeddings", "/v1/messages":
-		return true
-	default:
-		return false
-	}
-}
-
-func doJSONRequest(ctx context.Context, method, url, apiKey string, out any, client *http.Client) (int, http.Header, error) {
-	req, err := http.NewRequestWithContext(ctx, method, url, nil)
-	if err != nil {
-		return 0, nil, err
-	}
-	if apiKey != "" {
-		req.Header.Set("Authorization", "Bearer "+apiKey)
-	}
-	req.Header.Set("Accept", "application/json")
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return 0, nil, err
-	}
-	defer resp.Body.Close()
-
-	if out == nil {
-		io.Copy(io.Discard, resp.Body)
-		return resp.StatusCode, resp.Header, nil
-	}
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return resp.StatusCode, resp.Header, err
-	}
-	if len(body) == 0 {
-		return resp.StatusCode, resp.Header, nil
-	}
-	if err := json.Unmarshal(body, out); err != nil {
-		return resp.StatusCode, resp.Header, err
-	}
-	return resp.StatusCode, resp.Header, nil
-}
-
-func doJSONPostRequest(ctx context.Context, url string, body any, out any, client *http.Client) (int, error) {
-	payload, err := json.Marshal(body)
-	if err != nil {
-		return 0, err
-	}
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, url, bytes.NewReader(payload))
-	if err != nil {
-		return 0, err
-	}
-	req.Header.Set("Accept", "application/json")
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return 0, err
-	}
-	defer resp.Body.Close()
-
-	if out == nil {
-		io.Copy(io.Discard, resp.Body)
-		return resp.StatusCode, nil
-	}
-
-	respBody, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return resp.StatusCode, err
-	}
-	if len(respBody) == 0 {
-		return resp.StatusCode, nil
-	}
-	if err := json.Unmarshal(respBody, out); err != nil {
-		return resp.StatusCode, err
-	}
-	return resp.StatusCode, nil
-}
-
-func sanitizeMetricPart(input string) string {
-	s := strings.ToLower(strings.TrimSpace(input))
-	s = nonAlnumRe.ReplaceAllString(s, "_")
-	s = strings.Trim(s, "_")
-	if s == "" {
-		return "unknown"
-	}
-	return s
-}
-
-func normalizeModelName(input string) string {
-	s := strings.TrimSpace(strings.ToLower(input))
-	if s == "" {
-		return ""
-	}
-	s = strings.Trim(strings.TrimPrefix(s, "models/"), "/")
-	if strings.HasPrefix(s, "ollama.com/") {
-		s = strings.TrimPrefix(s, "ollama.com/")
-	}
-	if i := strings.LastIndex(s, "/"); i >= 0 {
-		s = s[i+1:]
-	}
-	s = strings.TrimSpace(strings.TrimSuffix(s, ":latest"))
-	return s
-}
-
-func cloudEndpointURL(base, path string) string {
-	base = strings.TrimRight(strings.TrimSpace(base), "/")
-	if base == "" {
-		base = defaultCloudBaseURL
-	}
-	if !strings.HasPrefix(path, "/") {
-		path = "/" + path
-	}
-	return base + path
-}
-
-func resolveCloudSessionCookie(acct core.AccountConfig) string {
-	if acct.ExtraData != nil {
-		for _, key := range []string{"cloud_session_cookie", "session_cookie", "cookie"} {
-			if v := strings.TrimSpace(acct.ExtraData[key]); v != "" {
-				return v
-			}
-		}
-	}
-	if v := strings.TrimSpace(os.Getenv("OLLAMA_SESSION_COOKIE")); v != "" {
-		return v
-	}
-	return ""
-}
-
-func fetchCloudUsageFromSettingsPage(ctx context.Context, cloudBaseURL, apiKey string, acct core.AccountConfig, snap *core.UsageSnapshot, client *http.Client) (bool, error) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, cloudEndpointURL(cloudBaseURL, "/settings"), nil)
-	if err != nil {
-		return false, fmt.Errorf("ollama: creating settings request: %w", err)
-	}
-	req.Header.Set("Accept", "text/html,application/xhtml+xml")
-	if apiKey != "" {
-		req.Header.Set("Authorization", "Bearer "+apiKey)
-	}
-	if cookie := resolveCloudSessionCookie(acct); cookie != "" {
-		req.Header.Set("Cookie", cookie)
-	}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return false, fmt.Errorf("ollama: cloud settings request failed: %w", err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode == http.StatusUnauthorized || resp.StatusCode == http.StatusForbidden {
-		return false, nil
-	}
-	if resp.StatusCode != http.StatusOK {
-		return false, fmt.Errorf("ollama: cloud settings endpoint returned HTTP %d", resp.StatusCode)
-	}
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return false, fmt.Errorf("ollama: reading cloud settings response: %w", err)
-	}
-
-	pcts := make(map[string]float64)
-	for _, m := range settingsUsageRe.FindAllStringSubmatch(string(body), -1) {
-		if len(m) < 3 {
-			continue
-		}
-		label := strings.ToLower(strings.TrimSpace(m[1]))
-		v, convErr := strconv.ParseFloat(strings.TrimSpace(m[2]), 64)
-		if convErr != nil {
-			continue
-		}
-		pcts[label] = v
-	}
-
-	resets := make(map[string]time.Time)
-	for _, m := range settingsResetRe.FindAllStringSubmatch(string(body), -1) {
-		if len(m) < 3 {
-			continue
-		}
-		label := strings.ToLower(strings.TrimSpace(m[1]))
-		t, ok := parseAnyTime(strings.TrimSpace(m[2]))
-		if !ok {
-			continue
-		}
-		resets[label] = t
-	}
-
-	found := false
-	if v, ok := pcts["session usage"]; ok {
-		snap.Metrics["usage_five_hour"] = core.Metric{
-			Used:   core.Float64Ptr(v),
-			Unit:   "%",
-			Window: "5h",
-		}
-		if t, ok := resets["session usage"]; ok {
-			snap.Resets["usage_five_hour"] = t
-			snap.SetAttribute("block_end", t.Format(time.RFC3339))
-			snap.SetAttribute("block_start", t.Add(-5*time.Hour).Format(time.RFC3339))
-		}
-		found = true
-	}
-	if v, ok := pcts["weekly usage"]; ok {
-		weekly := core.Metric{
-			Used:   core.Float64Ptr(v),
-			Unit:   "%",
-			Window: "1w",
-		}
-		snap.Metrics["usage_weekly"] = weekly
-		// Backward-compatible alias for existing widgets/config.
-		snap.Metrics["usage_one_day"] = core.Metric{
-			Used:   core.Float64Ptr(v),
-			Unit:   "%",
-			Window: "1d",
-		}
-		if t, ok := resets["weekly usage"]; ok {
-			snap.Resets["usage_weekly"] = t
-			snap.Resets["usage_one_day"] = t
-		}
-		found = true
-	}
-
-	return found, nil
-}
-
-func setValueMetric(snap *core.UsageSnapshot, key string, value float64, unit, window string) {
-	snap.Metrics[key] = core.Metric{
-		Used:      core.Float64Ptr(value),
-		Remaining: core.Float64Ptr(value),
-		Unit:      unit,
-		Window:    window,
-	}
-}
-
-func fileExists(path string) bool {
-	info, err := os.Stat(path)
-	return err == nil && !info.IsDir()
-}
-
-func summarizeModels(models []tagModel, limit int) string {
-	if len(models) == 0 || limit <= 0 {
-		return ""
-	}
-	out := make([]string, 0, limit)
-	for i := 0; i < len(models) && i < limit; i++ {
-		name := normalizeModelName(models[i].Name)
-		if name == "" {
-			name = normalizeModelName(models[i].Model)
-		}
-		if name == "" {
-			continue
-		}
-		out = append(out, name)
-	}
-	return strings.Join(out, ", ")
-}
-
-func normalizeHeaderKey(k string) string {
-	return strings.ReplaceAll(strings.ToLower(strings.TrimSpace(k)), "-", "_")
-}
-
-func isCloudModel(model tagModel) bool {
-	name := strings.ToLower(strings.TrimSpace(model.Name))
-	mdl := strings.ToLower(strings.TrimSpace(model.Model))
-	if strings.HasSuffix(name, ":cloud") || strings.HasSuffix(mdl, ":cloud") {
-		return true
-	}
-	if strings.TrimSpace(model.RemoteHost) != "" || strings.TrimSpace(model.RemoteModel) != "" {
-		return true
-	}
-	return false
-}
-
-func anyValueCaseInsensitive(m map[string]any, keys ...string) (any, bool) {
-	if len(m) == 0 {
-		return nil, false
-	}
-	want := make(map[string]struct{}, len(keys))
-	for _, key := range keys {
-		norm := normalizeLookupKey(key)
-		if norm == "" {
-			continue
-		}
-		want[norm] = struct{}{}
-	}
-	for k, v := range m {
-		if _, ok := want[normalizeLookupKey(k)]; ok {
-			return v, true
-		}
-	}
-	return nil, false
-}
-
-func anyStringCaseInsensitive(m map[string]any, keys ...string) string {
-	v, ok := anyValueCaseInsensitive(m, keys...)
-	if !ok {
-		return ""
-	}
-	switch val := v.(type) {
-	case string:
-		return strings.TrimSpace(val)
-	case fmt.Stringer:
-		return strings.TrimSpace(val.String())
-	default:
-		return ""
-	}
-}
-
-func anyMapCaseInsensitive(m map[string]any, keys ...string) map[string]any {
-	v, ok := anyValueCaseInsensitive(m, keys...)
-	if !ok {
-		return nil
-	}
-	out, _ := v.(map[string]any)
-	return out
-}
-
-func anyBoolCaseInsensitive(m map[string]any, keys ...string) (bool, bool) {
-	v, ok := anyValueCaseInsensitive(m, keys...)
-	if !ok {
-		return false, false
-	}
-	switch val := v.(type) {
-	case bool:
-		return val, true
-	case string:
-		b, err := strconv.ParseBool(strings.TrimSpace(val))
-		if err == nil {
-			return b, true
-		}
-	}
-	return false, false
-}
-
-func anyFloatCaseInsensitive(m map[string]any, keys ...string) (float64, bool) {
-	v, ok := anyValueCaseInsensitive(m, keys...)
-	if !ok {
-		return 0, false
-	}
-	return anyFloat(v)
-}
-
-func anyFloat(v any) (float64, bool) {
-	switch val := v.(type) {
-	case float64:
-		return val, true
-	case float32:
-		return float64(val), true
-	case int:
-		return float64(val), true
-	case int64:
-		return float64(val), true
-	case int32:
-		return float64(val), true
-	case uint:
-		return float64(val), true
-	case uint64:
-		return float64(val), true
-	case uint32:
-		return float64(val), true
-	case json.Number:
-		f, err := val.Float64()
-		if err == nil {
-			return f, true
-		}
-	case string:
-		s := strings.TrimSpace(strings.TrimSuffix(val, "%"))
-		f, err := strconv.ParseFloat(s, 64)
-		if err == nil {
-			return f, true
-		}
-	}
-	return 0, false
-}
-
-func anyNullStringCaseInsensitive(m map[string]any, keys ...string) string {
-	raw := anyMapCaseInsensitive(m, keys...)
-	if len(raw) == 0 {
-		return ""
-	}
-	valid, ok := anyBoolCaseInsensitive(raw, "valid")
-	if ok && !valid {
-		return ""
-	}
-	return anyStringCaseInsensitive(raw, "string", "value")
-}
-
-func anyNullTimeCaseInsensitive(m map[string]any, keys ...string) (time.Time, bool) {
-	raw := anyMapCaseInsensitive(m, keys...)
-	if len(raw) == 0 {
-		return time.Time{}, false
-	}
-	valid, ok := anyBoolCaseInsensitive(raw, "valid")
-	if ok && !valid {
-		return time.Time{}, false
-	}
-	timeRaw := anyStringCaseInsensitive(raw, "time", "value")
-	if timeRaw == "" {
-		return time.Time{}, false
-	}
-	return parseAnyTime(timeRaw)
-}
-
-func normalizeLookupKey(s string) string {
-	s = strings.TrimSpace(strings.ToLower(s))
-	s = strings.ReplaceAll(s, "_", "")
-	s = strings.ReplaceAll(s, "-", "")
-	s = strings.ReplaceAll(s, ".", "")
-	return s
-}
-
-func parseAnyTime(raw string) (time.Time, bool) {
-	t, err := shared.ParseTimestampString(raw)
-	if err != nil {
-		return time.Time{}, false
-	}
-	return t, true
 }
 
 type versionResponse struct {
